Visual representation
A Potentiometer that controls three leds. When the Potentiometer is turned to it’s lowest state, the blue led will turn on, with the red and green lights turned off. If the Potentiometer is turned, the blue light will turn off, and the green lights brightness will adjust with the Potentiometer. When the Potentiometer is turned into its highest state, the red light will turn on, and the green light will be at its highest brightness.